requirements.txt
flet==0.22.*
pycryptodome